+++ src/main/java/io/github/astrapi69/collections/array/ArrayFactory.java
@@ -1,8 +1,8 @@
 /**
  * The MIT License
- *
+ * <p>
  * Copyright (C) 2015 Asterios Raptis
- *
+ * <p>
  * Permission is hereby granted, free of charge, to any person obtaining
  * a copy of this software and associated documentation files (the
  * "Software"), to deal in the Software without restriction, including
@@ -10,10 +10,10 @@
  * distribute, sublicense, and/or sell copies of the Software, and to
  * permit persons to whom the Software is furnished to do so, subject to
  * the following conditions:
- *
+ * <p>
  * The above copyright notice and this permission notice shall be
  * included in all copies or substantial portions of the Software.
- *
+ * <p>
  * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND,
  * EXPRESS OR IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF
  * MERCHANTABILITY, FITNESS FOR A PARTICULAR PURPOSE AND
@@ -24,6 +24,9 @@
  */
 package io.github.astrapi69.collections.array;
 
+import java.io.ByteArrayOutputStream;
+import java.io.DataOutputStream;
+import java.io.IOException;
 import java.lang.reflect.Array;
 
 import io.github.astrapi69.check.Argument;
@@ -101,6 +104,30 @@ public final class ArrayFactory
 	public static byte[] newByteArray(final byte... elements)
 	{
 		return elements;
+	}
+
+	/**
+	 * Factory method for create new array of primitive byte values from the given optional
+	 * elements.
+	 *
+	 * @param elements
+	 *            the optional primitive byte values that will be in the returned array of primitive
+	 *            byte values array.
+	 * @return the new array of primitive byte values
+	 */
+	@SafeVarargs
+	public static byte[] newByteArray(final int... elements)
+	{
+		byte[] array = new byte[0];
+		if (elements != null && 0 < elements.length)
+		{
+			array = new byte[elements.length];
+			for (int i = 0; i < elements.length; ++i)
+			{
+				array[i] = (byte) elements[i];
+			}
+		}
+		return array;
 	}
 
 	/**